Move bulk of compilation logic to own module
